renames the util file to something descriptive
